PythonBackend/combined_app/requirements.txt
Flask
flask-cors
pymongo
numpy
pandas
scikit-learn
xgboost
statsmodels
prophet
face-recognition
opencv-python
python-dotenv
PyJWT
Werkzeug